refactor: replace default text
