+++ spring-ai-alibaba-playground/ui/src/App.tsx
@@ -114,7 +114,7 @@ const Independent: React.FC = () => {
       <div className={styles.layout}>
         <FunctionMenu />
         {/* 菜单页面容器 */}
-        <div style={{ flex: 1, position: "relative", overflow: "hidden" }}>
+        <div className={styles.menuPagesWrapper}>
           <Routes>
             <Route path="/" element={<Navigate to="/chat" replace />} />
             {Object.entries(pageComponents).map(([key, Component]) => (
@@ -123,18 +123,7 @@ const Independent: React.FC = () => {
                 <Route
                   path={`/${key}`}
                   element={
-                    <div
-                      style={{
-                        position: "absolute",
-                        top: 0,
-                        left: 0,
-                        width: "100%",
-                        height: "100%",
-                        transition: "none",
-                        backgroundColor: token.colorBgContainer,
-                        overflowY: "auto",
-                      }}
-                    >
+                    <div className={styles.pageWrapper}>
                       <Component />
                     </div>
                   }
@@ -143,18 +132,7 @@ const Independent: React.FC = () => {
                 <Route
                   path={`/${key}/:conversationId`}
                   element={
-                    <div
-                      style={{
-                        position: "absolute",
-                        top: 0,
-                        left: 0,
-                        width: "100%",
-                        height: "100%",
-                        transition: "none",
-                        backgroundColor: token.colorBgContainer,
-                        overflowY: "auto",
-                      }}
-                    >
+                    <div className={styles.pageWrapper}>
                       <Component />
                     </div>
                   }
@@ -164,7 +142,6 @@ const Independent: React.FC = () => {
           </Routes>
         </div>
 
-        {/* 页脚始终显示 */}
         <Layout.Footer
           className={styles.footer}
           style={{ color: token.colorText }}
