+++ spring-ai-alibaba-playground/ui/src/menuPages/components/FunctionMenu/index.tsx
@@ -8,13 +8,13 @@ import {
 } from "@ant-design/icons";
 import { Button, message, Select, Space, Typography, Input } from "antd";
 import React, { useEffect, useState, useRef } from "react";
-import { useStyle } from "../../../style";
+import { useStyle } from "./style";
 import { useModelConfigContext } from "../../../stores/modelConfig.store";
 import {
   Conversation,
   useConversationContext,
 } from "../../../stores/conversation.store";
-import { functionMenuItems } from "../../../const";
+import { functionMenuItems } from "./const";
 import { useFunctionMenuStore } from "../../../stores/functionMenu.store";
 import { useNavigate } from "react-router-dom";
 
@@ -23,9 +23,7 @@ export interface ConversationItem {
   label: React.ReactNode;
 }
 
-export interface MenuProps {}
-
-const FunctionMenu = (props: MenuProps) => {
+const FunctionMenu = () => {
   const { styles } = useStyle();
   const { menuCollapsed, toggleMenuCollapsed } = useFunctionMenuStore();
   const {
@@ -88,7 +86,7 @@ const FunctionMenu = (props: MenuProps) => {
     navigate("/chat");
   };
 
-  // 开始编辑会话标题
+  // 编辑会话标题
   const startEditingTitle = (
     conversation: Conversation,
     e: React.MouseEvent
@@ -96,7 +94,7 @@ const FunctionMenu = (props: MenuProps) => {
     e.stopPropagation();
     setEditingConversationId(conversation.id);
     setEditingTitle(conversation.title);
-    // 等待DOM更新后聚焦输入框
+    // 等待 DOM 更新后聚焦输入框
     setTimeout(() => {
       if (inputRef.current) {
         inputRef.current.focus();
